* Merge pgbouncer_1_8_1 into master branch
* Merge upstream: 8b604da5 into master branch

+++ test/test_json.c
@@ -1,4 +1,3 @@
-
 #include <usual/json.h>
 
 #include <usual/string.h>
@@ -31,7 +30,7 @@ static const char *simple_value(const char *json)
 		if (!json_value_as_int(obj, &val)) {
 			res = "ELONG";
 		} else {
-			snprintf(buf, sizeof(buf), "INT:%lld", (long long)val);
+			snprintf(buf, sizeof(buf), "INT:%" PRId64, val);
 			res = buf;
 		}
 	} else if (json_value_is_float(obj)) {
@@ -432,4 +431,3 @@ struct testcase_t json_tests[] = {
 	{ "relax", test_json_relax },
 	END_OF_TESTCASES
 };
-
